simulação de exibição em egressos
