--- src/pages/Egressos/Egressos.js
@@ -1,16 +1,0 @@
-import React from 'react'
-import Header from '../../components/Header'
-import Title from '../../components/Title'
-
-const Egressos = () => {
-    return(
-        <React.Fragment>
-            <Header></Header>
-            <div className={'egressosPg'}>
-                <Title name={"Egressos"}/> 
-            </div>
-        </React.Fragment>
-    )
-}
-
-export default Egressos
