+++ src/pages/Egressos/fakeData.json
@@ -1,0 +1,23 @@
+[
+  {
+      "nome": "Iana Daya",
+      "ano_conclusao": "1986",
+      "id": 1,
+      "trabalho_atual": "Professora"
+  },
+  {
+    "nome": "Iana Maria",
+    "ano_conclusao": "1988",
+    "id": 2,
+    "trabalho_atual": "Desenvolvedora"
+      
+  },
+  {
+    "nome": "Iana Alexandra",
+    "ano_conclusao": "1989",
+    "id": 3,
+    "trabalho_atual": "Pesquisadora"
+  }
+  
+
+]
